(CU-85zt0qwa2 all) refactor: folders setup for props and ui
