+++ addon/ui/__init__.py
@@ -26,45 +26,3 @@
 # see <http://www.gnu.org/licenses/>.
 #
 # ##### END LICENSE BLOCK #####
-
-class BM_LABELS_Props():
-    labels = {
-        'Global': {
-            "short_bake_instruction": {
-                "description": "Press `ESC` key to cancel baking current map iteration.\n\nOpen Blender Console to see more information about the baking process and, if you face an unexpected Blender freeze, be able to press `Ctrl + C` (Windows), `Cmd + C` (Mac), `Super + C` (Linux) to abort the bake. Enable Prompt before freeze for more control",  # noqa: E501
-            }
-        }
-    }
-
-    def __init__(self, property_group: str, property_name: str,
-                 property_arg: str):
-        self.property_group = property_group
-        self.property_name = property_name
-        self.property_arg = property_arg
-
-    def get(self):
-        try:
-            label = self.labels[self.property_group][self.property_name][
-                self.property_arg]
-        except KeyError:
-            print("asdfsadf")
-            label = ""
-        return label
-
-
-class BM_LABELS_Operators():
-    labels = {
-    }
-
-    def __init__(self, operator_name: str, property_name: str):
-        self.operator_name = operator_name
-        self.property_name = property_name
-
-    def get(self):
-        try:
-            label = self.labels[self.operator_name][self.property_name]
-        except KeyError:
-            label = ""
-        return label
-
-
